Latest part IV
Here goes the last part IV. I used parts II and III adjusted, just removed the part that dropped lat and long. I created the tables and took the normal test. Not good (normality), although the dataset is large. Just to be sure, tried an elastic net without transformation. R2 0.74, to the 0.77 transformed. It is there.